fixed slopeTo rounding problem

+++ week-3/collinear/src/Point.java
@@ -69,7 +69,7 @@ public class Point implements Comparable<Point> {
             return Double.POSITIVE_INFINITY;
         if (this.y == that.y)
             return +0.0;
-        return (this.y - that.y)/(this.x - that.x);
+        return ((double)(this.y - that.y)) / ((double)(this.x - that.x));
     }
 
     /**
